requirements.txt
Flask==2.3.2
line-bot-sdk==2.4.2
pip==20.2.4
pydub==0.25.1
python-dotenv==1.0.0
requests==2.28.2
SpeechRecognition==3.10.0

